+++ doc/source/changelog.rst
@@ -47,6 +47,7 @@ CMake
 
   * added openssl, libevent_openssl
   * rename dbus into dbus-1, and dbus-glib into dbus-glib-1
+  * ``qi_use_lib(OPENGL)`` now uses upstream's ``FindOpenGL.cmake``
 
 * added :cmake:function:`qi_generate_src`,  :cmake:function:`qi_generate_header`
 * :cmake:function:`qi_create_gtest` and :cmake:function:`qi_create_test` now